refactor: refactor delete tiles
